Move the sample app to away from the main code and with docs
