+++ .gitignore
@@ -18,3 +18,4 @@
 vendor/
 webhook-broker
 dist/
+docs/tech-specs/basic-spec-spikes/basic-spec-spikes
