+++ docs/tech-specs/basic-spec-spikes/go.mod
@@ -1,0 +1,5 @@
+module github.com/imyousuf/webhook-broker/docs/tech-specs/basic-spec-spikes
+
+go 1.15
+
+require github.com/google/wire v0.4.0
