+++ docs/tech-specs/basic-spec-spikes/processor.go
