+++ docs/tech-specs/basic-spec-spikes/wire.go
