+++ docs/tech-specs/basic-spec-spikes/wire_gen.go
